更新一键动行.jar

+++ script/run-jar.sh
@@ -2,14 +2,20 @@
 #!/bin/sh
 
 # docker-compose.yml
-echo 'curl -sOL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/docker-compose.yml'
-curl -sOL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/docker-compose.yml
-if [ ! -d config ];then
-  mkdir config
+if [ ! -f ./docker-compose.yml ];then
+  echo 'curl -sOL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/docker-compose.yml'
+  curl -sOL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/docker-compose.yml
 fi
-echo 'curl -sL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/config/application-dev.yml -o ./config/application-dev.yml'
-curl -sL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/config/application-dev.yml -o ./config/application-dev.yml
+
+if [ ! -f ./config/application-dev.yml ];then
+  if [ ! -d config ];then
+    mkdir config
+  fi
+  echo 'curl -sL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/config/application-dev.yml -o ./config/application-dev.yml'
+  curl -sL https://gitee.com/smallsaas/sandbox/raw/master/tag/run/jar/config/application-dev.yml -o ./config/application-dev.yml
+fi
 
 ## start
 echo "docker-compose up"
+echo "docker-compose up -d"
 docker-compose up
